Moved SavedScanDetail links into their own view.
